air-quality-ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "air-quality-ml"
version = "0.1.0"
description = "Spark MLlib + MLflow pipeline for air quality forecasting and alerting"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "AI Coding Agent" }
]
dependencies = [
  "pyspark==3.5.1",
  "mlflow==2.17.2",
  "pandas==2.2.3",
  "numpy==2.1.2",
  "pydantic==2.10.6",
  "PyYAML==6.0.2",
  "scikit-learn==1.5.2",
  "scipy==1.14.1",
  "plotly==5.24.1",
  "pymongo==4.10.1",
  "prometheus-client==0.21.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
